+++ im_flutter_sdk_android/android/src/main/java/com/easemob/im_flutter_sdk/MethodKey.java
@@ -8,7 +8,6 @@ public class MethodKey {
     static final String logout = "logout";
     static final String changeAppKey = "changeAppKey";
 
-    static final String uploadLog = "uploadLog";
     static final String compressLogs = "compressLogs";
     static final String kickDevice = "kickDevice";
     static final String kickAllDevices = "kickAllDevices";
@@ -130,7 +129,6 @@ public class MethodKey {
     static final String loadMsgWithMsgType = "loadMsgWithMsgType";
     static final String loadMsgWithTime = "loadMsgWithTime";
     static final String messageCount = "messageCount";
-    static final String removeMsgFromServerWithTimeStamp = "removeMsgFromServerWithTimeStamp";
 
     // EMMessage method
     static final String getReactionList = "getReactionList";
@@ -227,13 +225,9 @@ public class MethodKey {
     static final String onGroupChanged = "onGroupChanged";
 
     /// EMPushManager
-    static final String getImPushConfig = "getImPushConfig";
     static final String getImPushConfigFromServer = "getImPushConfigFromServer";
     static final String updateImPushStyle = "updateImPushStyle";
     static final String updatePushNickname = "updatePushNickname";
-    static final String updateHMSPushToken = "updateHMSPushToken";
-    static final String updateFCMPushToken = "updateFCMPushToken";
-    static final String reportPushAction = "reportPushAction";
     static final String setConversationSilentMode = "setConversationSilentMode";
     static final String removeConversationSilentMode = "removeConversationSilentMode";
     static final String fetchConversationSilentMode = "fetchConversationSilentMode";
@@ -247,9 +241,7 @@ public class MethodKey {
 
     /// EMUserInfoManager 
     static final String updateOwnUserInfo = "updateOwnUserInfo";
-    static final String updateOwnUserInfoWithType = "updateOwnUserInfoWithType";
     static final String fetchUserInfoById = "fetchUserInfoById";
-    static final String fetchUserInfoByIdWithType = "fetchUserInfoByIdWithType";
 
     /// EMPresenceManager methods
     static final String presenceWithDescription = "publishPresenceWithDescription";